+++ PivotView.Core/Layout/PivotLayout.cs
@@ -4,22 +4,22 @@ abstract public class PivotLayout
 {
     public virtual int ItemCountOverride { get; set; }
 
-    public virtual double ItemAspectRatioOverride { get; set; }
+    public virtual float ItemAspectRatioOverride { get; set; }
 
-    public virtual double ItemMargin { get; set; }
+    public virtual float ItemMargin { get; set; }
 
     /// <summary>
     /// Width/Height
     /// </summary>
-    public double ItemAspectRatio { get; protected set; } = 1.0;
+    public float ItemAspectRatio { get; protected set; } = 1.0f;
 
-    public double ItemWidth { get; protected set; }
+    public float ItemWidth { get; protected set; }
 
-    public double ItemHeight { get; protected set; }
+    public float ItemHeight { get; protected set; }
 
-    public double LayoutWidth { get; protected set; }
+    public float LayoutWidth { get; protected set; }
 
-    public double LayoutHeight { get; protected set; }
+    public float LayoutHeight { get; protected set; }
 
     public void LayoutItems(IReadOnlyList<PivotRendererItem> items, RectangleF frame)
     {
@@ -31,15 +31,15 @@ abstract public class PivotLayout
 
     public abstract void ArrangeItems(IReadOnlyList<PivotRendererItem> items, RectangleF frame);
 
-    protected double GetItemAspectRatio(IReadOnlyList<PivotRendererItem> items)
+    protected float GetItemAspectRatio(IReadOnlyList<PivotRendererItem> items)
     {
-        if (ItemAspectRatioOverride != 0.0)
+        if (ItemAspectRatioOverride != 0.0f)
             return ItemAspectRatioOverride;
 
         // just use the first item for now
         if (items.Count > 0)
             return items[0].AspectRatio;
 
-        return 1.0;
+        return 1.0f;
     }
 }
